+++ model/Product.js
@@ -19,7 +19,7 @@ const Product =
         },
         price: {
             allowNull: false,
-            type: Sequelize.FLOAT
+            type: Sequelize.DECIMAL(10, 2)
         }
     });
 export {Product};
